purge unnecessary converting

+++ app/src/main/java/com/suihan74/satena/scenes/entries2/EntriesActivity.kt
@@ -684,7 +684,7 @@ class EntriesActivity : AppCompatActivity(), ScrollableToTop {
             override fun onTransitionCompleted(motionLayout: MotionLayout?, currentId: Int) {
                 binding.mainLayout.updatePadding(
                     top = when (currentId) {
-                        R.id.end -> extraMargin.toInt()
+                        R.id.end -> extraMargin
                         else -> 0
                     }
                 )
